split content-script functionality to modules
