+++ manifest.json
@@ -1,6 +1,6 @@
 {
   "manifest_version": 3,
-  "name": "LA-armory-enhance",
+  "name": "LA-armory-ext",
   "version": "0.0.2",
   "action": {
     "default_icon": {
