+++ src/constants/NegativeEngravings.js
@@ -1,0 +1,1 @@
+export const NEGATIVE_ENGRAVINGS = ['Износ', 'Ржавчина', 'Медлительность', 'Миролюбие'];
