+++ src/constants/SettingsOptions.js
@@ -1,0 +1,5 @@
+export const SETTINGS_DEFAULT = {
+    loadAtStart: true,
+    modifyEngravings: true,
+    displayQuality: true
+}
