+++ src/content-scripts/modules/gs-personal.js
@@ -1,0 +1,7 @@
+export function loadAllCharacters(characterList) {
+    characterList.forEach(char => {
+        chrome.runtime.sendMessage({type: 'load-character-info', name: char.name}, (response) => {
+            char.element.innerHTML = char.name + '(' + response.gs + ')'
+        });
+    })
+}
