+++ vite.config.js
@@ -6,9 +6,6 @@ import manifest from './manifest.json' assert {type: 'json'} // Node >=17
 
 // https://vitejs.dev/config/
 export default defineConfig({
-    build: {
-        target: 'esnext'
-    },
     plugins: [
         vue(),
         crx({manifest}),
